implement mipmap

+++ Core/Resource/src/TextureManager.cpp
@@ -15,7 +15,7 @@ VulkanTexture TextureManager::CreateResource(const std::string &key, const Sampl
     int                  height = 0;
     const unsigned char *data   = file_system::LoadTexture(key, &width, &height);
 
-    return VulkanTexture(width, height, VK_FORMAT_R8G8B8A8_UNORM, sizeof(unsigned char) * 4, data, config);
+    return VulkanTexture(static_cast<uint32_t>(width), static_cast<uint32_t>(height), VK_FORMAT_R8G8B8A8_UNORM, sizeof(unsigned char) * 4, data, config);
 }
 
 void TextureManager::Init() {
